@@ -1,5 +1,5 @@
 {
-  "name": "node-esm",
+  "name": "next-webpack5",
   "version": "1.0.0",
   "license": "MIT",
   "scripts": {
@@ -7,16 +7,16 @@
   },
   "dependencies": {
     "@emotion/core": "11.0.0",
-    "@emotion/react": "11.7.1",
-    "@emotion/styled": "11.6.0",
-    "@mui/material": "5.0.0-alpha.34",
-    "@mui/icons-material": "5.0.0-alpha.34",
-    "@mui/lab": "5.0.0-alpha.34",
-    "@mui/styled-engine": "5.0.0-alpha.34",
-    "@mui/styles": "5.0.0-alpha.33",
-    "@mui/system": "5.0.0-alpha.34",
-    "@mui/utils": "5.0.0-alpha.33",
-    "next": "12.1.0",
+    "@emotion/react": "11.10.4",
+    "@emotion/styled": "11.10.4",
+    "@mui/material": "5.10.5",
+    "@mui/icons-material": "5.10.3",
+    "@mui/lab": "5.0.0-alpha.99",
+    "@mui/styled-engine": "5.10.5",
+    "@mui/styles": "5.10.3",
+    "@mui/system": "5.10.5",
+    "@mui/utils": "5.10.3",
+    "next": "12.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-is": "18.2.0"
